+++ js/MFF.js
@@ -3,7 +3,7 @@
 // localStorage.removeItem("sorter");
 var MFF =
 {
- "version" : "2.2.2",
+ "version" : "2.3.0",
  "versionMFF" : "2.9.5",
  "localStorageKey" : "characters",
  "toid" : null,
@@ -76,6 +76,10 @@ var MFF =
  "axisItems" :
  {
   "name" :             { "disableChart" : true, "label" : "Name", "callback" : function(character) { return { "value" : MFF.CHARACTERS.DATA[character.id].uniforms[character.uniform].name, "percent" : false }; } },
+  "completion" :       { "label" : "% global", "callback" : function(character) { var v = MFF.PERCENT.get(character.id); return { "value" : v, "percent" : true }; } },
+  "percent_gears" :    { "label" : "% gears", "callback" : function(character) { var v = MFF.PERCENT.getGears(character.id); return { "value" : v, "percent" : true }; } },
+  "percent_skills" :   { "label" : "% skills", "callback" : function(character) { var v = MFF.PERCENT.getSkills(character.id); return { "value" : v, "percent" : true }; } },
+  "percent_uniform" :  { "label" : "% uniform", "callback" : function(character) { var v = MFF.PERCENT.getUniform(character.id); return { "value" : v, "percent" : true }; } },
   "attack" :           { "label" : "Attack", "callback" : function(character) { var v = character.attack[MFF.CHARACTERS.DATA[character.id].uniforms[character.uniform].attackBase]; return { "value" : v, "percent" : false }; } },
   "attack_energy" :    { "label" : "Energy attack", "callback" : function(character) { var v = character.attack.energy; return { "value" : v, "percent" : false }; } },
   "attack_physical" :  { "label" : "Physical attack", "callback" : function(character) { var v = character.attack.physical; return { "value" : v, "percent" : false }; } },
@@ -85,22 +89,20 @@ var MFF =
   "critrate" :         { "label" : "Critical rate", "max" : 75, "callback" : function(character) { var v = character.critrate; return { "value" : v, "percent" : true }; } },
   // TODO : need max debuff duration
   "debuff" :           { "label" : "Debuff duration", "callback" : function(character) { var v = character.debuff; return { "value" : v, "percent" : true }; } },
-  "completion" :       { "label" : "Development completion", "callback" : function(character) { var v = MFF.computePercent(character.id); return { "value" : v, "percent" : true }; } },
   "defense_average" :  { "label" : "Defense average", "callback" : function(character) { var v = (character.defense.physical + character.defense.energy) / 2; return { "value" : v, "percent" : false }; } },
   "defense_energy" :   { "label" : "Defense energy", "callback" : function(character) { var v = character.defense.energy; return { "value" : v, "percent" : false }; } },
   "defense_physical" : { "label" : "Defense physical", "callback" : function(character) { var v = character.defense.physical; return { "value" : v, "percent" : false }; } },
-  // TODO : need max dodge
-  "dodge" :            { "label" : "Dodge", "callback" : function(character) { var v = character.dodge; return { "value" : v, "percent" : true }; } },
+  "dodge" :            { "label" : "Dodge", "max" : 75, "callback" : function(character) { var v = character.dodge; return { "value" : v, "percent" : true }; } },
   "hp" :               { "label" : "Hit points", "callback" : function(character) { var v = character.hp; return { "value" : v, "percent" : false }; } },
   "defpen" :           { "label" : "Ignore defense", "max" : 50, "callback" : function(character) { var v = character.defpen; return { "value" : v, "percent" : true }; } },
-  // TODO : need max ignore dodge
-  "ignore_dodge" :     { "label" : "Ignore dodge", "callback" : function(character) { var v = character.ignore_dodge; return { "value" : v, "percent" : true }; } },
+  // TODO : check if max ignore dodge is really 75
+  "ignore_dodge" :     { "label" : "Ignore dodge", "max" : 75, "callback" : function(character) { var v = character.ignore_dodge; return { "value" : v, "percent" : true }; } },
   "last_update" :      { "disableChart" : true, "label" : "Last update", "callback" : function(character) { return { "value" : character.lastUpdate || null, "percent" : false }; } },
   "level" :            { "label" : "Level", "callback" : function(character) { var v = character.level; return { "value" : v, "percent" : false }; } },
   "movspeed" :         { "label" : "Movement speed", "max" : 130, "callback" : function(character) { var v = character.movspeed; return { "value" : v, "percent" : true }; } },
+  "rank" :             { "label" : "Rank", "callback" : function(character) { var v = character.rank || 0; return { "value" : v, "percent" : false }; } },
   "recorate" :         { "label" : "Recovery rate", "max" : 250, "callback" : function(character) { var v = character.recorate; return { "value" : v, "percent" : true }; } },
-  // toDO : need max skill cooldown
-  "scd" :              { "label" : "Skill cooldown", "callback" : function(character) { var v = character.scd; return { "value" : v, "percent" : true }; } },
+  "scd" :              { "label" : "Skill cooldown", "max" : 50, "callback" : function(character) { var v = character.scd; return { "value" : v, "percent" : true }; } },
   "defense_all" :      { "disableCharts" : true, "disableSort" : true, "label" : "All defenses" },
   "attack_all" :       { "disableCharts" : true, "disableSort" : true, "label" : "All attacks" }
  },
@@ -168,56 +170,66 @@ var MFF =
   /* compatibility 2.1.0 end */
   Highcharts.setOptions({ "plotOptions" : { "series" : { "animation" : false } } });
   MFF.LAYOUT.init();
+  API.EVT.on("refreshPercentGlobal", function(character)
+                                     {
+                                      var div = document.getElementById("current_percent"),
+                                          v = MFF.PERCENT.get(character, true);
+                                      if ( div ) { div.innerHTML = v + "%"; }
+                                     });
  },
- "getIndividualPercent" : function(cur, min, max)
- {
-  return cur > 0 ? API.numberToFixed(((cur - min) / (max - min)) * 100, 2) : 0;
- },
- "computePercent" : function(character)
- {
-  var i, j,
-      total = 0,
-      maxi = 0,
-      data = MFF.CHARACTERS.get(character);
-  for ( i = 0; i < data.gear.length; i++ )
-  {
-   for ( j = 0; j < data.gear[i].length; j++ )
-   {
-    maxi += 100;
-    if ( data.gear[i][j].pref ) { total += data.gear[i][j].percent; }
-   }
-  }
-  return total * 100 / maxi;
- },
+ // "getIndividualPercent" : function(cur, min, max)
+ // {
+ //  return cur > 0 ? API.numberToFixed(((cur - min) / (max - min)) * 100, 2) : 0;
+ // },
+ // "computePercent" : function(character)
+ // {
+ //  return 0;
+ //  /*
+ //  var i, j,
+ //      total = 0,
+ //      maxi = 0,
+ //      data = MFF.CHARACTERS.get(character);
+ //  for ( i = 0; i < data.gear.length; i++ )
+ //  {
+ //   for ( j = 0; j < data.gear[i].length; j++ )
+ //   {
+ //    maxi += 100;
+ //    if ( data.gear[i][j].pref ) { total += data.gear[i][j].percent; }
+ //   }
+ //  }
+ //  return total * 100 / maxi;
+ //  */
+ // },
  "toid2" : null,
  "saveCharacter" : function(data)
  {
-  MFF.CHARACTERS.setProperty(MFF.currentCharacter, data);
+  MFF.CHARACTERS.setProperty(MFF.currentCharacter || MFF.lastTarget, data);
   localStorage.setItem(MFF.localStorageKey, JSON.stringify(MFF.CHARACTERS.getAll()));
   if ( data.mode == "tier" )
   {
    API.EVT.dispatch("updateTier");
    MFF.LAYOUT.LIST.setTier(MFF.currentCharacter);
   }
-  else if ( data.mode == "gear" || data.mode == "skill" )
-  {
-   MFF.toid2 = clearTimeout(MFF.toid2);
-   MFF.toid2 = setTimeout(function()
-                          {
-                           var div = document.getElementById("current_percent");
-                           if ( div && MFF.currentCharacter )
-                           {
-                            div.innerHTML = API.numberToFixed(MFF.computePercent(MFF.currentCharacter), 2) + "%";
-                            API.EVT.dispatch("updateTier");
-                            if ( MFF.LAYOUT.DETAIL.GEARS._btnDetailCharts.isActive() )
-                            {
-                             MFF.LAYOUT.CHARTS.renderDetail();
-                            }
-                            MFF.LAYOUT.LIST.setSub(MFF.currentCharacter);
-                            MFF.LAYOUT.LIST.synchroDetailGear(MFF.currentCharacter);
-                           }
-                          }, 250);
-  }
+//   else if ( data.mode == "gear" || data.mode == "skill" )
+//   {
+//    MFF.toid2 = clearTimeout(MFF.toid2);
+//    MFF.toid2 = setTimeout(function()
+//                           {
+//                            var div = document.getElementById("current_percent");
+//                            if ( div && MFF.currentCharacter )
+//                            {
+// //                            div.innerHTML = API.numberToFixed(MFF.computePercent(MFF.currentCharacter), 2) + "%";
+//                             div.innerHTML = MFF.PERCENT.get(MFF.currentCharacter, true) + "%";
+//                             API.EVT.dispatch("updateTier");
+//                             if ( MFF.LAYOUT.DETAIL.GEARS._btnDetailCharts.isActive() )
+//                             {
+//                              MFF.LAYOUT.CHARTS.renderDetail();
+//                             }
+//                             MFF.LAYOUT.LIST.setSub(MFF.currentCharacter);
+//                             MFF.LAYOUT.LIST.synchroDetailGear(MFF.currentCharacter);
+//                            }
+//                           }, 250);
+//   }
   else if ( data.mode == "uniform" )
   {
    MFF.LAYOUT.DETAIL.drawCharacter(MFF.currentCharacter, true, true);